Import Recipes from recipes.json
In order to import the recipes from recipes.json, we need to create
at least one database table. We already know that we will want to
search by ingredients, so let's make it completely relational and
save each ingredient in it's own row (VS in a JSON column in `recipes`
table). Same for tags, since we don't know yet how they will be used.

So we have 3 tables: recipes, ingredients and tags, and their related
ActiveRecord model.

The import rake task takes ~237 seconds to import locally: not bad, not
terrible. Since this is init phase and nothing is production ready yet,
this is acceptable.